pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "utilskit"
version = "0.2.18"
description = "다양한 편의성 함수 모음 패키지"
requires-python = ">=3.10"
authors = [
    { name = "kimyh", email = "kim_yh663927@naver.com" },
]
dependencies = [
    "matplotlib==3.10.3",
    "numpy==2.2.6",
    "pandas==2.3.1",
    "PyMySQL==1.1.1",
    "SQLAlchemy==2.0.41",
    "tqdm==4.67.1",
    "xlrd==2.0.2",
    "openpyxl==3.1.5",
    "semver>=3.0.4",
    "toml-w",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.optional-dependencies]
plot = [
    "matplotlib",
    "seaborn",
]
all = [
    "matplotlib",
    "scikit-learn",
    "numpy",
    "pandas",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [
    ".",
]
exclude = [
    "test/",
    "scripts/",
    "old/",
    "*.tmp",
]

[tool.setuptools.package-data]
